restructuring files and folders
